Scripting Fundamentals
As I never ever had to learn scripting for my undergrad course, my current skills are basically none in this matter. Then, reading from and writing to files should be a good place to start and, in particular, I bought this course on web developing with python which will help me on the journey.